+++ myMY/idGen.py
@@ -1,0 +1,2 @@
+""" unique id generator """
+from uuid import *